moved token request logic to its context file
